refactor: split user totp routes
